Extract some code into `escaping.rs`
